daemon: plugin poc
This is a POC demonstration for an internal plugin system that allows
for in-tree plugins. Contributors can drop `pkg/daemon/internal/plugin`
packages that will be able to used all the internal MCD functionality,
but MCD will not be able to use them.

HERE BE DRAGONS: this is a POC, so its _really_ rough